refactor(reco-core): projections own their CPU surface maps
stitch/geometry.rs (PlaneMap + l_shape_plane_maps) moves into
projection/l_shape.rs; stitch/cylinder.rs (CylinderMap) moves into
projection/cylinder.rs. Pure moves: each file had exactly one consumer,
the projection that owns the math. Former module docs fold into the
map types' docs; visibility tightens to module-private. stitch/ keeps
the executor, the compositor primitives and the oracle.
